fix: Avoiding converting a JSON object with circular references
A JSON object is said to have a circular reference when it contains a reference to itself. This can happen when an object is nested within itself, or when two or more objects reference each other. When you try to serialize a JSON object with circular references using JSON.stringify(), you'll get a TypeError because the method can't handle circular references.